+++ kedro_kubeflow/generator.py
@@ -97,7 +97,7 @@ class PipelineGenerator(object):
         image,
         image_pull_policy,
     ) -> Dict[str, dsl.ContainerOp]:
-        """Build kfp container graph from Kedro node dependencies. """
+        """Build kfp container graph from Kedro node dependencies."""
         kfp_ops = {}
 
         node_volumes = (
@@ -169,6 +169,7 @@ class PipelineGenerator(object):
                         + self.catalog[output]["filepath"]
                         for output in node.outputs
                         if output in self.catalog
+                        and "filepath" in self.catalog[output]
                     },
                 ),
                 image_pull_policy,