+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Increased test coverage for the CLI handling
+
 ## [0.1.10] - 2021-01-11
 
 ### Added