+++ CHANGELOG.md
@@ -2,10 +2,19 @@
 
 ## [Unreleased]
 
+### Added 
+
+- Created initial documentation hosted on kedro-kubeflow.readthedocs.io
+
+### Fixed
+
+- Fixed broken `kubeflow init` command (#29)
+
 ### Changed
 
 - Increased test coverage for the CLI handling
 - Ability to change the effective user id for steps if the ownership of the volume needs it
+
 
 ## [0.1.10] - 2021-01-11